chore: repo structure refactored

+++ install_dependencies.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
 
 # install apt packages
-sudo apt install -y $(cat apt-packages.txt)
+if type apt &>/dev/null; then
+  sudo apt install -y $(cat apt-packages.txt)
+else
+  echo "This does not appear to be an ubuntu VM."
+  exit 1
+fi
 
 # install bat via deb package as it isn't available in Debian 10
-wget https://github.com/sharkdp/bat/releases/download/v0.15.4/bat_0.15.4_amd64.deb
-sudo dpkg -i ./bat_0.15.4_amd64.deb
-rm bat_0.15.4_amd64.deb
+if ! type bat &>/dev/null; then
+  wget https://github.com/sharkdp/bat/releases/download/v0.15.4/bat_0.15.4_amd64.deb
+  sudo dpkg -i ./bat_0.15.4_amd64.deb
+  rm bat_0.15.4_amd64.deb
+fi
 
 # install python dependencies
-python3.7 -m pip install -r requirements.txt
+if type uv &>/dev/null; then
+  uv sync
+else
+  python3.9 -m pip install -r requirements.txt
+fi
